+++ FaunaDB.Client.Test/TestCase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
+using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using FaunaDB.Client;
@@ -17,6 +18,7 @@ namespace Test
     public class TestCase
     {
         protected static Field<string> SECRET_FIELD = Field.At("secret").To<string>();
+        protected static Field<RefV> REF_FIELD = Field.At("ref").To<RefV>();
         protected const string testDbName = "faunadb-csharp-test";
 
         protected FaunaClient rootClient;
@@ -27,6 +29,9 @@ namespace Test
         protected Value adminKey;
         protected String faunaEndpoint;
         protected String faunaSecret;
+        
+        protected RefV GetRef(Value v) =>
+            v.Get(REF_FIELD);
 
         [OneTimeSetUp]
         public void SetUp()
@@ -79,6 +84,27 @@ namespace Test
         {
             var resp = new RequestResult(HttpMethodKind.Get, "", null, "", responseText, (int)statusCode, null, DateTime.UtcNow, DateTime.UtcNow);
             return new FaunaClient(clientIO: new MockClientIO(resp));
+        }
+        
+        protected async Task<RefV> RandomCollection()
+        {
+            Value coll = await client.Query(
+                CreateCollection(
+                    Obj("name", RandomStartingWith("some_coll_")))
+            );
+
+            return GetRef(coll);
+        }
+        
+        protected string RandomStartingWith(params string[] strs)
+        {
+            StringBuilder builder = new StringBuilder();
+            foreach (var str in strs)
+                builder.Append(str);
+
+            builder.Append(new Random().Next(0, int.MaxValue));
+
+            return builder.ToString();
         }
     }
